Simplify dashboard pages by removing complex interactions and charts
